+++ CHANGELOG.md
@@ -6,6 +6,7 @@
  - Melhoria: [Implementado **Hover Inspect** no debug](https://github.com/totvs/advpl-vscode/issues/417)
  - Melhoria: [Permitir **excluir um ambiente pelo gerenciador de ambientes**](https://github.com/totvs/advpl-vscode/issues/408)
  - Melhoria: [Permitir excluir fontes via **menu Explorer**](https://github.com/totvs/advpl-vscode/issues/445)
+ - Melhoria: [Adicionar opção no **Env Manager** para trocar a senha](https://github.com/totvs/advpl-vscode/issues/448)
  - Melhoria: Permitir cancelar a compilação de arquivos abertos
  - Correção: [Ao compilar arquivos abertos, compilava a Workspace inteira](https://github.com/totvs/advpl-vscode/issues/411)
  - Correção: [JSON Bridge Return Invalid](https://github.com/totvs/advpl-vscode/issues/409)